+++ modules/flopy/modflow/mfhob.py
@@ -586,7 +586,7 @@ class HeadObservation(object):
 
         # convert passed time_series_data to a numpy array
         if isinstance(time_series_data, list):
-            time_series_data = np.array(time_series_data, dtype=np.float)
+            time_series_data = np.array(time_series_data, dtype=np.float64)
 
         # if a single observation is passed as a list reshape to a
         # two-dimensional numpy array
@@ -672,7 +672,7 @@ class HeadObservation(object):
 
         """
         # get the default HOB dtype
-        dtype = np.dtype([("totim", np.float32), ("irefsp", np.int),
-                          ("toffset", np.float32),
-                          ("hobs", np.float32), ("obsname", '|S12')])
+        dtype = np.dtype([("totim", np.float64), ("irefsp", np.int_),
+                          ("toffset", np.float64),
+                          ("hobs", np.float64), ("obsname", '|S12')])
         return dtype
